+++ kinto_http/cli_utils.py
@@ -34,8 +34,8 @@ class AuthAction(argparse.Action):
         if values is not None:
             auth = None
             try:
-                columnIndex = values.find(":")
-                if columnIndex == -1 or values.index(" ") < columnIndex:
+                colonIndex = values.find(":")
+                if colonIndex == -1 or values.index(" ") < colonIndex:
                     # Handle: `username:password with spaces` versus `Bearer TokenWith:Semicolon`
                     bearer_type, bearer_token = values.split(" ", 1)
                     auth = BearerTokenAuth(token=bearer_token, type=bearer_type)